ImageResizer compile module now resizes only block and item textures
Game was crasing because ImageResizer was resizing all png files. That
caused game to crash because some files were smaller then they should
be. This commit fixes that by checking if file that is currently
compiling comes from either one of directories:

assets/minecraft/textures/blocks
assets/minecraft/textures/items

If is the files from different directory, it is now simply skipped in
ImageResizer. 

There is a plan to make more configurable which files should and which
shouldn't be resized in future.